CMS增加商户代金券配置板块 实现逻辑删除 ALTER TABLE `ny_merchant_card` ADD COLUMN `deleted` TINYINT(1) DEFAULT 0 COMMENT '是否删除';

+++ src/main/java/com/nuanyou/cms/entity/enums/CardTypeConverter.java
@@ -10,7 +10,7 @@ public class CardTypeConverter implements AttributeConverter<CardType, Integer> 
     public Integer convertToDatabaseColumn(CardType attribute) {
         if (attribute == null)
             return null;
-        return attribute.value.intValue();
+        return attribute.value;
     }
 
     @Override
